+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/FileSystemsTest.java
@@ -116,6 +116,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import com.oracle.truffle.api.TruffleFile;
+import com.oracle.truffle.api.TruffleFile.FileStoreInfo;
 import com.oracle.truffle.api.TruffleLanguage.Env;
 import com.oracle.truffle.api.TruffleLanguage.Registration;
 import com.oracle.truffle.api.nodes.RootNode;
@@ -159,7 +160,7 @@ public class FileSystemsTest {
     private static final String MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES = "Memory file system with language homes";
     private static final String MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES_INTERNAL_FILE = "Memory file system with language homes - internal file";
     private static final String CONTEXT_PRE_INITIALIZATION_FILESYSTEM_BUILD_TIME = "Context pre-initialization file system build time";
-    private static final String CONTEXT_PRE_INITIALIZATION_FILESYSTEM_EXECUTION_TIME = "Context pre-initialization files ystem execution time";
+    private static final String CONTEXT_PRE_INITIALIZATION_FILESYSTEM_EXECUTION_TIME = "Context pre-initialization file system execution time";
     private static final String COMPOSITE_FILE_SYSTEM_DELEGATE = "Composite file system read write delegate";
     private static final String COMPOSITE_FILE_SYSTEM_FALLBACK = "Composite file system read only fallback";
     private static final String DENY_IO = "Deny IO file system";
@@ -222,7 +223,7 @@ public class FileSystemsTest {
         ctx = Context.newBuilder().engine(engine).allowIO(IOAccess.NONE).build();
         setCwd(ctx, privateDir);
         cfgs.put(NO_IO_UNDER_LANGUAGE_HOME_PUBLIC_FILE,
-                        new Configuration(NO_IO_UNDER_LANGUAGE_HOME_PUBLIC_FILE, ctx, privateDir, privateDir, fullIO, true, false, false, false, false, true, engine));
+                        new Configuration(NO_IO_UNDER_LANGUAGE_HOME_PUBLIC_FILE, ctx, privateDir, privateDir, fullIO, true, false, false, false, false, false, true, engine));
 
         // No IO under language home - internal file
         engine = Engine.create();
@@ -232,7 +233,7 @@ public class FileSystemsTest {
         ctx = Context.newBuilder().engine(engine).allowIO(IOAccess.NONE).build();
         setCwd(ctx, privateDir);
         cfgs.put(NO_IO_UNDER_LANGUAGE_HOME_INTERNAL_FILE,
-                        new Configuration(NO_IO_UNDER_LANGUAGE_HOME_INTERNAL_FILE, ctx, privateDir, privateDir, fullIO, true, true, false, false, true, false, engine));
+                        new Configuration(NO_IO_UNDER_LANGUAGE_HOME_INTERNAL_FILE, ctx, privateDir, privateDir, fullIO, true, true, false, false, true, false, false, engine));
 
         // Read Only
         IOAccess ioAccess;
@@ -294,17 +295,16 @@ public class FileSystemsTest {
         cfgs.put(MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES, new Configuration(MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES, ctx, memDir, fileSystem, false, true, true, true));
 
         // Memory with language home - in language home
-        fileSystem = FileSystem.allowInternalResourceAccess(new MemoryFileSystem());
-        memDir = mkdirs(fileSystem.toAbsolutePath(fileSystem.parsePath("work")), fileSystem);
-        fileSystem.setCurrentWorkingDirectory(memDir);
-        privateDir = createContent(memDir, fileSystem);
+        privateDir = createContent(Files.createTempDirectory(FileSystemsTest.class.getSimpleName()).toRealPath(), fullIO);
         engine = Engine.create();
         markAsLanguageHome(engine, SetCurrentWorkingDirectoryLanguage.class, privateDir);
+        fileSystem = FileSystem.allowInternalResourceAccess(new MemoryFileSystem());
+        fileSystem.setCurrentWorkingDirectory(privateDir);
         ioAccess = IOAccess.newBuilder().fileSystem(fileSystem).build();
         ctx = Context.newBuilder().engine(engine).allowIO(ioAccess).build();
         setCwd(ctx, privateDir);
         cfgs.put(MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES_INTERNAL_FILE,
-                        new Configuration(MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES_INTERNAL_FILE, ctx, privateDir, privateDir, fileSystem, false, true, true, true, true, true, engine));
+                        new Configuration(MEMORY_FILE_SYSTEM_WITH_LANGUAGE_HOMES_INTERNAL_FILE, ctx, privateDir, privateDir, fileSystem, false, true, true, true, true, true, true, engine));
 
         // PreInitializeContextFileSystem in image build time
         fileSystem = createPreInitializeContextFileSystem();
@@ -2712,6 +2712,50 @@ public class FileSystemsTest {
         }
     }
 
+    @Test
+    public void testGetFileStoreInfo() {
+        Context ctx = cfg.getContext();
+        String configuration = cfg.getName();
+        String path = cfg.getPath().toString();
+        boolean usePublicFile = cfg.usePublicFile;
+        boolean canRead = cfg.canReadFileStoreInfo();
+        AbstractExecutableTestLanguage.evalTestLanguage(ctx, TestGetFileStoreInfoLanguage.class, "", configuration, path, usePublicFile, canRead);
+    }
+
+    @Registration
+    public static final class TestGetFileStoreInfoLanguage extends AbstractExecutableTestLanguage {
+
+        @Override
+        @TruffleBoundary
+        protected Object execute(RootNode node, Env env, Object[] contextArguments, Object[] frameArguments) throws Exception {
+            String configurationName = (String) contextArguments[0];
+            String path = (String) contextArguments[1];
+            boolean usePublicFile = (boolean) contextArguments[2];
+            boolean canRead = (boolean) contextArguments[3];
+            TruffleFile file = resolve(env, usePublicFile, path, FOLDER_EXISTING, FILE_EXISTING);
+            FileStoreInfo fsInfo = file.getFileStoreInfo();
+            try {
+                fsInfo.getTotalSpace();
+                Assert.assertTrue(formatErrorMessage("Expected SecurityException", configurationName, path), canRead);
+            } catch (SecurityException se) {
+                Assert.assertFalse(formatErrorMessage("Unexpected SecurityException", configurationName, path), canRead);
+            }
+            try {
+                fsInfo.getUnallocatedSpace();
+                Assert.assertTrue(formatErrorMessage("Expected SecurityException", configurationName, path), canRead);
+            } catch (SecurityException se) {
+                Assert.assertFalse(formatErrorMessage("Unexpected SecurityException", configurationName, path), canRead);
+            }
+            try {
+                fsInfo.getUsableSpace();
+                Assert.assertTrue(formatErrorMessage("Expected SecurityException", configurationName, path), canRead);
+            } catch (SecurityException se) {
+                Assert.assertFalse(formatErrorMessage("Unexpected SecurityException", configurationName, path), canRead);
+            }
+            return null;
+        }
+    }
+
     static TruffleFile resolve(Env env, boolean usePublicFile, String path, String... paths) {
         TruffleFile file;
         if (usePublicFile) {
@@ -2812,6 +2856,7 @@ public class FileSystemsTest {
         private final boolean writableFileSystem;
         private final boolean allowsUserDir;
         private final boolean allowsAbsolutePath;
+        private final boolean allowsFileStoreInfo;
         private final boolean usePublicFile;
         private final Engine engine;
 
@@ -2837,11 +2882,11 @@ public class FileSystemsTest {
                         final boolean readableFileSystem,
                         final boolean writableFileSystem,
                         final boolean allowsUserDir) {
-            this(name, context, path, userDir, fileSystem, internalFileSystem, readableFileSystem, writableFileSystem, allowsUserDir, allowsUserDir, true, null);
+            this(name, context, path, userDir, fileSystem, internalFileSystem, readableFileSystem, writableFileSystem, allowsUserDir, allowsUserDir, readableFileSystem, true, null);
         }
 
         Configuration(String name, Context context, Path path, Path userDir, FileSystem fileSystem, boolean internalFileSystem, boolean readableFileSystem,
-                        boolean writableFileSystem, boolean allowsUserDir, boolean allowsAbsolutePath, boolean usePublicFile, Engine engine) {
+                        boolean writableFileSystem, boolean allowsUserDir, boolean allowsAbsolutePath, boolean allowsFileStoreInfo, boolean usePublicFile, Engine engine) {
             Objects.requireNonNull(name, "Name must be non null.");
             Objects.requireNonNull(context, "Context must be non null.");
             Objects.requireNonNull(path, "Path must be non null.");
@@ -2857,6 +2902,7 @@ public class FileSystemsTest {
             this.writableFileSystem = writableFileSystem;
             this.allowsUserDir = allowsUserDir;
             this.allowsAbsolutePath = allowsAbsolutePath;
+            this.allowsFileStoreInfo = allowsFileStoreInfo;
             this.usePublicFile = usePublicFile;
             this.engine = engine;
         }
@@ -2894,6 +2940,14 @@ public class FileSystemsTest {
          */
         boolean canRead() {
             return readableFileSystem;
+        }
+
+        /**
+         * Returns true if the test configuration allows reading of file store info, such as free
+         * disk space.
+         */
+        boolean canReadFileStoreInfo() {
+            return allowsFileStoreInfo;
         }
 
         /**
@@ -3071,7 +3125,12 @@ public class FileSystemsTest {
 
     static void markAsLanguageHome(Engine engine, Class<? extends TruffleLanguage<?>> language, Path languageHome) {
         try (Context ctx = Context.newBuilder().engine(engine).build()) {
-            AbstractExecutableTestLanguage.evalTestLanguage(ctx, MarkAsLanguageHomeLanguage.class, "", TestUtils.getDefaultLanguageId(language), languageHome.toString());
+            String id = TestUtils.getDefaultLanguageId(language);
+            String home = languageHome.toString();
+            // Set language home in isolate
+            AbstractExecutableTestLanguage.evalTestLanguage(ctx, MarkAsLanguageHomeLanguage.class, "", id, home);
+            // Set language home in host
+            MarkAsLanguageHomeLanguage.markAsLanguageHome(id, home);
         }
     }
 
@@ -3081,10 +3140,14 @@ public class FileSystemsTest {
         @TruffleBoundary
         protected Object execute(RootNode node, Env env, Object[] contextArguments, Object[] frameArguments) throws Exception {
             String languageId = (String) contextArguments[0];
-            String langHome = (String) contextArguments[1];
-            System.setProperty("org.graalvm.language." + languageId + ".home", langHome);
-            resetLanguageHomes();
+            String languageHome = (String) contextArguments[1];
+            markAsLanguageHome(languageId, languageHome);
             return null;
+        }
+
+        static void markAsLanguageHome(String languageId, String languageHome) {
+            System.setProperty("org.graalvm.language." + languageId + ".home", languageHome);
+            resetLanguageHomes();
         }
     }
 
@@ -3227,6 +3290,21 @@ public class FileSystemsTest {
         public boolean isSameFile(Path path1, Path path2, LinkOption... options) throws IOException {
             return delegate.isSameFile(path1, path2, options);
         }
+
+        @Override
+        public long getTotalSpace(Path path) throws IOException {
+            return delegate.getTotalSpace(path);
+        }
+
+        @Override
+        public long getUnallocatedSpace(Path path) throws IOException {
+            return delegate.getUnallocatedSpace(path);
+        }
+
+        @Override
+        public long getUsableSpace(Path path) throws IOException {
+            return delegate.getUsableSpace(path);
+        }
     }
 
     private static final class RestrictedFileSystem extends ForwardingFileSystem {
@@ -3327,6 +3405,24 @@ public class FileSystemsTest {
         public void setAttribute(Path path, String attribute, Object value, LinkOption... options) throws IOException {
             checkWrite(path);
             super.setAttribute(path, attribute, value, options);
+        }
+
+        @Override
+        public long getTotalSpace(Path path) throws IOException {
+            checkRead(path);
+            return super.getTotalSpace(path);
+        }
+
+        @Override
+        public long getUnallocatedSpace(Path path) throws IOException {
+            checkRead(path);
+            return super.getUnallocatedSpace(path);
+        }
+
+        @Override
+        public long getUsableSpace(Path path) throws IOException {
+            checkRead(path);
+            return super.getUsableSpace(path);
         }
 
         private void checkRead(Path path) {